# docs/swagger.yaml
basePath: /
definitions:
...
  consts.CardType:
    enum:
    - 1
    - 2
    - 3
    type: integer
    x-enum-varnames:
    - StudentCard
    - NormalCard
    - VipCard
  consts.Role:
    enum:
    - 1
    - 2
    - 3
    type: integer
    x-enum-varnames:
    - AdminRole
    - UserRole
    - StaffRole
  consts.Status:
    enum:
    - active
    - inactive
    - blocked
    type: string
    x-enum-varnames:
    - ActiveStatus
    - InactiveStatus
    - BlockedStatus
...
  models.Card:
    properties:
      balance:
        type: number
      created_at:
        type: string
      id:
        type: integer
      price:
        type: number
      rf_id:
        type: string
      status:
        $ref: '#/definitions/consts.Status'
      type:
        $ref: '#/definitions/consts.CardType'
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: quan hệ belongs-to
      username:
        description: 1-1 nên unique
        type: string
    type: object
...
  models.Station:
    properties:
      created_at:
        type: string
      id:
        type: integer
      ip_address:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.StationHistory:
    properties:
      action:
        description: '"checkin" hoặc "checkout"'
        type: string
      card:
        allOf:
        - $ref: '#/definitions/models.Card'
        description: Foreign key relationships
      card_id:
        type: string
      created_at:
        type: string
      id:
        type: integer
      station:
        $ref: '#/definitions/models.Station'
      station_id:
        type: integer
      time:
        type: string
      updated_at:
        type: string
      used_balance:
        type: number
    type: object
  models.Train:
    properties:
      company:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      role:
        allOf:
        - $ref: '#/definitions/consts.Role'
        description: '"ADMIN", "USER", "STAFF"'
      status:
        description: '"active" hoặc "inactive"'
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A Metro Card Management System API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Metro API
  version: "1.0"
paths:
...
  /station-history:
    get:
      consumes:
      - application/json
      description: Retrieve all station history records with optional filtering
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of records per page
        in: query
        name: limit
        type: integer
      - description: Filter by card ID
        in: query
        name: card_id
        type: string
      - description: Filter by station ID
        in: query
        name: station_id
        type: integer
      - description: Filter by action (checkin/checkout)
        in: query
        name: action
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Station histories retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.StationHistory'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get all station history records
      tags:
      - station-history
  /station-history/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a specific station history record by its ID
      parameters:
      - description: Station History ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Station history retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.StationHistory'
              type: object
        "404":
          description: Station history not found
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get station history by ID
      tags:
      - station-history
  /station-history/action/{action}:
    get:
      consumes:
      - application/json
      description: Retrieve all station history records for a specific action (checkin/checkout)
      parameters:
      - description: Action (checkin/checkout)
        in: path
        name: action
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Station histories retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.StationHistory'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get station histories by action
      tags:
      - station-history
  /station-history/card/{card_id}:
    get:
      consumes:
      - application/json
      description: Retrieve all station history records for a specific card
      parameters:
      - description: Card ID
        in: path
        name: card_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Station histories retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.StationHistory'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get station histories by card ID
      tags:
      - station-history
  /station-history/station/{station_id}:
    get:
      consumes:
      - application/json
      description: Retrieve all station history records for a specific station
      parameters:
      - description: Station ID
        in: path
        name: station_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Station histories retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.StationHistory'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get station histories by station ID
      tags:
      - station-history
...
  /train:
    get:
      consumes:
      - application/json
      description: Retrieve all train records with optional filtering
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of records per page
        in: query
        name: limit
        type: integer
      - description: Filter by train type
        in: query
        name: type
        type: string
      - description: Filter by company
        in: query
        name: company
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Trains retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Train'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get all train records
      tags:
      - train
    post:
      consumes:
      - application/json
      description: Create a new train record
      parameters:
      - description: Train information
        in: body
        name: train
        required: true
        schema:
          $ref: '#/definitions/models.Train'
      produces:
      - application/json
      responses:
        "200":
          description: Train created successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Train'
              type: object
        "400":
          description: Bad request - validation error
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Create a new train record
      tags:
      - train
  /train/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a train record
      parameters:
      - description: Train ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Train deleted successfully
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Train not found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Delete train
      tags:
      - train
    get:
      consumes:
      - application/json
      description: Retrieve a specific train record by its ID
      parameters:
      - description: Train ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Train retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Train'
              type: object
        "404":
          description: Train not found
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get train by ID
      tags:
      - train
    put:
      consumes:
      - application/json
      description: Update an existing train record
      parameters:
      - description: Train ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated train information
        in: body
        name: train
        required: true
        schema:
          $ref: '#/definitions/models.Train'
      produces:
      - application/json
      responses:
        "200":
          description: Train updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Train'
              type: object
        "400":
          description: Bad request - validation error
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Train not found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Update train
      tags:
      - train
  /train/company/{company}:
    get:
      consumes:
      - application/json
      description: Retrieve all train records for a specific company
      parameters:
      - description: Company name
        in: path
        name: company
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Trains retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Train'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get trains by company
      tags:
      - train
  /train/type/{type}:
    get:
      consumes:
      - application/json
      description: Retrieve all train records for a specific type
      parameters:
      - description: Train type
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Trains retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Train'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get trains by type
      tags:
      - train
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"